requirements.txt
bidict==0.21.3
python-engineio==4.2.1
python-socketio==5.4.1
requests==2.26.0
